moved cython functions to subdirectory
