+++ pyspike/DiscreteFunc.py
@@ -198,7 +198,7 @@ class DiscreteFunc(object):
 
         # cython version
         try:
-            from cython_add import add_discrete_function_cython as \
+            from cython.cython_add import add_discrete_function_cython as \
                 add_discrete_function_impl
         except ImportError:
             print("Warning: add_discrete_function_cython not found. Make \
@@ -206,7 +206,7 @@ sure that PySpike is installed by running\n\
 'python setup.py build_ext --inplace'! \
 \n Falling back to slow python backend.")
             # use python backend
-            from python_backend import add_discrete_function_python as \
+            from cython.python_backend import add_discrete_function_python as \
                 add_discrete_function_impl
 
         self.x, self.y, self.mp = \
