+++ pyspike/PieceWiseConstFunc.py
@@ -146,14 +146,14 @@ class PieceWiseConstFunc(object):
 
         # cython version
         try:
-            from cython_add import add_piece_wise_const_cython as \
+            from cython.cython_add import add_piece_wise_const_cython as \
                 add_piece_wise_const_impl
         except ImportError:
             print("Warning: add_piece_wise_const_cython not found. Make sure \
 that PySpike is installed by running\n 'python setup.py build_ext --inplace'! \
 \n Falling back to slow python backend.")
             # use python backend
-            from python_backend import add_piece_wise_const_python as \
+            from cython.python_backend import add_piece_wise_const_python as \
                 add_piece_wise_const_impl
 
         self.x, self.y = add_piece_wise_const_impl(self.x, self.y, f.x, f.y)
