+++ pyspike/cython/cython_distance.pyx
