+++ setup.py
@@ -21,8 +21,8 @@ except ImportError:
 else:
     use_cython = True
 
-if os.path.isfile("pyspike/cython_add.c") and \
-   os.path.isfile("pyspike/cython_distance.c"):
+if os.path.isfile("pyspike/cython/cython_add.c") and \
+   os.path.isfile("pyspike/cython/cython_distance.c"):
     use_c = True
 else:
     use_c = False
@@ -32,14 +32,14 @@ ext_modules = []
 
 if use_cython:  # Cython is available, compile .pyx -> .c
     ext_modules += [
-        Extension("pyspike.cython_add", ["pyspike/cython_add.pyx"]),
-        Extension("pyspike.cython_distance", ["pyspike/cython_distance.pyx"]),
+        Extension("pyspike.cython.cython_add", ["pyspike/cython/cython_add.pyx"]),
+        Extension("pyspike.cython.cython_distance", ["pyspike/cython/cython_distance.pyx"]),
     ]
     cmdclass.update({'build_ext': build_ext})
 elif use_c:  # c files are there, compile to binaries
     ext_modules += [
-        Extension("pyspike.cython_add", ["pyspike/cython_add.c"]),
-        Extension("pyspike.cython_distance", ["pyspike/cython_distance.c"]),
+        Extension("pyspike.cython.cython_add", ["pyspike/cython/cython_add.c"]),
+        Extension("pyspike.cython.cython_distance", ["pyspike/cython/cython_distance.c"]),
     ]
 # neither cython nor c files available -> automatic fall-back to python backend
 
@@ -78,7 +78,7 @@ train similarity',
         'Programming Language :: Python :: 2.7',
     ],
     package_data={
-        'pyspike': ['cython_add.c', 'cython_distance.c'],
+        'pyspike': ['cython/cython_add.c', 'cython/cython_distance.c'],
         'test': ['Spike_testdata.txt']
     }
 )
